requirements.txt
django==6.0.2
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1
channels==4.3.2
channels-redis==4.3.0
psycopg2-binary==2.9.11
python-dotenv==1.2.1
redis==7.1.1



